+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Added
 - Created a tyre impression model with a status enum data type
 - Created database migration for new `tyre_impressions` table
+- Added a file service with file naming policies
 
 ## v0.0.4 21/05/2026
 ### Test Coverage